🔨 fix: event history decodeDangleToken로 role 전달
