+++ src/app/admin/shelter/event/page.tsx
@@ -1,61 +1,10 @@
-'use client';
-import useMyShelterEvent from '@/api/mypage/event/useMyShelterEvent';
-import useMyInfo from '@/api/mypage/useMyInfo';
-import { ChipOption } from '@/components/common/ChipInput/ChipInput';
-import EventHistory from '@/components/mypage/EventHistory/EventHistory';
-import useEventFilter from '@/components/mypage/EventHistory/hooks/useEventFilter';
-import useEventScroll from '@/components/mypage/EventHistory/hooks/useEventScroll';
-import { isShelterInfo } from '@/components/mypage/MyPageMain/MyPageMain';
-import useHeader from '@/hooks/useHeader';
-import { useAuthContext } from '@/providers/AuthContext';
-import { palette } from '@/styles/color';
+import MyShelterEventPage from '@/components/mypage/EventHistory/MyShelterEventPage';
+import { COOKIE_ACCESS_TOKEN_KEY } from '@/constants/cookieKeys';
+import decodeDangleToken from '@/utils/token/decodeDangleToken';
+import { cookies } from 'next/headers';
 
-export default function ShelterEvent() {
-  useHeader({ title: '봉사 활동 조회', color: palette.white });
-  const { dangle_role } = useAuthContext();
-  const { shelterFilter, handleChipInput } = useEventFilter();
-
-  const { data: info } = useMyInfo(dangle_role, {
-    enabled: !!dangle_role && dangle_role !== 'NONE'
-  });
-
-  const { data, fetchNextPage, hasNextPage, isLoading, isFetchingNextPage } =
-    useMyShelterEvent(shelterFilter);
-
-  useEventScroll({
-    isFetchingNextPage,
-    hasNextPage,
-    fetchNextPage,
-    shelterFilter
-  });
-
-  const STATUS_OPTIONS: ChipOption[] = [
-    {
-      label: `전체 ${
-        isShelterInfo(info)
-          ? info?.historyStat.done + info?.historyStat.inProgress
-          : 0
-      }`,
-      value: ''
-    },
-    {
-      label: `진행중 ${isShelterInfo(info) ? info?.historyStat.inProgress : 0}`,
-      value: 'IN_PROGRESS'
-    },
-    {
-      label: `종료 ${isShelterInfo(info) ? info?.historyStat.done : 0}`,
-      value: 'DONE'
-    }
-  ];
-
-  return (
-    <EventHistory
-      data={data!}
-      isLoading={isLoading}
-      isVolunteer={false}
-      shelterFilter={shelterFilter}
-      options={STATUS_OPTIONS}
-      onChange={handleChipInput}
-    />
-  );
+export default function ShelterHistoryPage() {
+  const accessToken = cookies().get(COOKIE_ACCESS_TOKEN_KEY)?.value || '';
+  const { dangle_role: role } = decodeDangleToken(accessToken);
+  return <MyShelterEventPage dangle_role={role} />;
 }
